+++ KeePassium/premium/LString+premium.swift
@@ -102,6 +102,10 @@ You can manage and cancel your subscriptions by going to your account settings o
         "[Premium/Price Plan/Conditions/future premium features]",
         value: "Future premium features",
         comment: "What's included in a price plan: upcoming premium features (as opposed to `current only`)")
+    public static let planConditionPerpetualFallback = NSLocalizedString(
+        "[Premium/Price Plan/Conditions/perpetual fallback]",
+        value: "Rent-to-own license (Perpetual fallback license)",
+        comment: "What's included in a price plan. Please leave `perpetual fallback license` in English.")
 
     
     public static let premiumCallToActionFree = NSLocalizedString(
@@ -151,5 +155,15 @@ You can manage and cancel your subscriptions by going to your account settings o
         value: "No previous purchase could be restored.",
         comment: "Text of an error message: there were no in-app purchases that can be restored")
     
+
+    public static let perpetualLicense = NSLocalizedString(
+        "[Premium/perpetualLicense]",
+        value: "Perpetual license",
+        comment: "Name of the permanent/never-ending license.")
+
+    public static let perpetualLicenseStatus = NSLocalizedString(
+        "[Premium/PerpetualLicense/status]",
+        value: "You have a perpetual license for all the versions before %@.",
+        comment: "[formattedDate: String]")
 
 }