+++ KeePassium/premium/PricingPlanCollectionCell.swift
@@ -48,7 +48,7 @@ class PricingPlanBenefitCell: UITableViewCell {
 
 protocol PricingPlanCollectionCellDelegate: class {
     func didPressPurchaseButton(in cell: PricingPlanCollectionCell, with pricePlan: PricingPlan)
-    func didPressPerpetualFallbackDetail(in cell: PricingPlanCollectionCell, with pricePlan: PricingPlan)
+    func didPressPerpetualFallbackInfo(in cell: PricingPlanConditionCell, with pricePlan: PricingPlan)
 }
 
 class PricingPlanCollectionCell: UICollectionViewCell {
@@ -323,6 +323,6 @@ extension PricingPlanCollectionCell: UITableViewDataSource {
 
 extension PricingPlanCollectionCell: PricingPlanConditionCellDelegate {
     func didPressDetailButton(in cell: PricingPlanConditionCell) {
-        delegate?.didPressPerpetualFallbackDetail(in: self, with: pricingPlan)
+        delegate?.didPressPerpetualFallbackInfo(in: cell, with: pricingPlan)
     }
 }