+++ KeePassium/premium/PricingPlans.swift
@@ -19,6 +19,7 @@ struct PricingPlanCondition {
         case allPremiumFeatures
         case upcomingPremiumFeatures
         case currentPremiumFeatures
+        case perpetualFallback
     }
     enum HelpReference {
         case none
@@ -46,6 +47,8 @@ struct PricingPlanCondition {
             return LString.planConditionUpcomingPremiumFeatures
         case .currentPremiumFeatures:
             return LString.planConditionCurrentPremiumFeatures
+        case .perpetualFallback:
+            return LString.planConditionPerpetualFallback
         }
     }
 }
@@ -240,6 +243,7 @@ class PricingPlanPremiumMonthly: RealPricingPlan {
             PricingPlanCondition(kind: .updatesAndFixes, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .emailSupport, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .allPremiumFeatures, isIncluded: true, moreInfo: .none),
+            PricingPlanCondition(kind: .perpetualFallback, isIncluded: true, moreInfo: .perpetualFallback),
         ]
         self.benefits = [
             PricingPlanBenefit.multipleDatabases,
@@ -266,6 +270,7 @@ class PricingPlanPremiumYearly: RealPricingPlan {
             PricingPlanCondition(kind: .updatesAndFixes, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .emailSupport, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .allPremiumFeatures, isIncluded: true, moreInfo: .none),
+            PricingPlanCondition(kind: .perpetualFallback, isIncluded: true, moreInfo: .perpetualFallback),
         ]
         self.benefits = [
             PricingPlanBenefit.multipleDatabases,