+++ KeePassium/settings/SettingsVC.swift
@@ -28,6 +28,7 @@ class SettingsVC: UITableViewController, Refreshable {
     @IBOutlet weak var premiumTrialCell: UITableViewCell!
     @IBOutlet weak var premiumStatusCell: UITableViewCell!
     @IBOutlet weak var manageSubscriptionCell: UITableViewCell!
+    @IBOutlet weak var appHistoryCell: UITableViewCell!
     
     private var settingsNotifications: SettingsNotifications!
     
@@ -38,6 +39,7 @@ class SettingsVC: UITableViewController, Refreshable {
         static let premiumTrial = IndexPath(row: 0, section: premiumSectionIndex)
         static let premiumStatus = IndexPath(row: 1, section: premiumSectionIndex)
         static let manageSubscription = IndexPath(row: 2, section: premiumSectionIndex)
+        static let appHistoryCell = IndexPath(row: 3, section: premiumSectionIndex)
     }
     private var hiddenIndexPaths = Set<IndexPath>()
     
@@ -69,6 +71,7 @@ class SettingsVC: UITableViewController, Refreshable {
             setPremiumCellVisibility(premiumTrialCell, isHidden: true)
             setPremiumCellVisibility(premiumStatusCell, isHidden: true)
             setPremiumCellVisibility(manageSubscriptionCell, isHidden: true)
+            setPremiumCellVisibility(appHistoryCell, isHidden: true)
         }
         refreshPremiumStatus()
         #if DEBUG
@@ -89,12 +92,14 @@ class SettingsVC: UITableViewController, Refreshable {
     
     deinit {
         appIconSwitcherCoordinator = nil
+        appHistoryCoordinator = nil
         premiumCoordinator = nil
     }
     
     func dismissPopover(animated: Bool) {
         self.dismiss(animated: animated) { [self] in 
             self.appIconSwitcherCoordinator = nil
+            self.appHistoryCoordinator = nil
             self.premiumCoordinator = nil
         }
     }
@@ -132,6 +137,8 @@ class SettingsVC: UITableViewController, Refreshable {
                 hiddenIndexPaths.insert(CellIndexPath.premiumStatus)
             case manageSubscriptionCell:
                 hiddenIndexPaths.insert(CellIndexPath.manageSubscription)
+            case appHistoryCell:
+                hiddenIndexPaths.insert(CellIndexPath.appHistoryCell)
             default:
                 break
             }
@@ -143,6 +150,8 @@ class SettingsVC: UITableViewController, Refreshable {
                 hiddenIndexPaths.remove(CellIndexPath.premiumStatus)
             case manageSubscriptionCell:
                 hiddenIndexPaths.remove(CellIndexPath.manageSubscription)
+            case appHistoryCell:
+                hiddenIndexPaths.remove(CellIndexPath.appHistoryCell)
             default:
                 break
             }
@@ -205,6 +214,8 @@ class SettingsVC: UITableViewController, Refreshable {
             didPressUpgradeToPremium()
         case manageSubscriptionCell:
             didPressManageSubscription()
+        case appHistoryCell:
+            didPressShowAppHistory()
         case diagnosticLogCell:
             let viewer = ViewDiagnosticsVC.make()
             show(viewer, sender: self)
@@ -271,6 +282,20 @@ class SettingsVC: UITableViewController, Refreshable {
             self?.appIconSwitcherCoordinator = nil
         }
         appIconSwitcherCoordinator!.start()
+    }
+    
+    var appHistoryCoordinator: AppHistoryCoordinator?
+    func didPressShowAppHistory() {
+        assert(appHistoryCoordinator == nil)
+        guard let navigationController = navigationController else {
+            fatalError()
+        }
+        let router = NavigationRouter(navigationController)
+        appHistoryCoordinator = AppHistoryCoordinator(router: router)
+        appHistoryCoordinator!.dismissHandler = { [weak self] coordinator in
+            self?.appHistoryCoordinator = nil
+        }
+        appHistoryCoordinator!.start()
     }