+++ KeePassium/util/LString.swift
@@ -312,6 +312,12 @@ public enum LString {
         comment: "Error message shown when the user tries to change file attributes (such as creation/modification date, backup flag, etc)"
     )
     
+
+    public static let titleHelpViewer = NSLocalizedString(
+        "[Help Viewer/title]",
+        value: "Help",
+        comment: "Generic title of in-app help article")
+    
     public static let titleTouchID  = NSLocalizedString(
         "[AppLock/Biometric/Hint] Unlock KeePassium",
         value: "Unlock KeePassium",