+++ KeePassiumLib/KeePassiumLib/db/Entry.swift
@@ -50,19 +50,20 @@ public class EntryField: Eraseable {
     public func contains(
         word: Substring,
         includeFieldNames: Bool,
-        includeProtectedValues: Bool
+        includeProtectedValues: Bool,
+        options: String.CompareOptions
     ) -> Bool {
         guard name != EntryField.password else { return false } 
         
         if includeFieldNames
             && !isStandardField
-            && name.localizedCaseInsensitiveContains(word)
+            && name.localizedContains(word, options: options)
         {
             return true
         }
         
         let includeFieldValue = !isProtected || includeProtectedValues
-        if includeFieldValue && value.localizedCaseInsensitiveContains(word) {
+        if includeFieldValue && value.localizedContains(word, options: options) {
             return true
         }
         return false
@@ -271,7 +272,8 @@ public class Entry: DatabaseItem, Eraseable {
                 wordFound = field.contains(
                     word: word,
                     includeFieldNames: query.includeFieldNames,
-                    includeProtectedValues: query.includeProtectedValues)
+                    includeProtectedValues: query.includeProtectedValues,
+                    options: query.compareOptions)
                 if wordFound {
                     break
                 }
@@ -281,7 +283,7 @@ public class Entry: DatabaseItem, Eraseable {
             }
 
             for att in attachments {
-                if att.name.localizedCaseInsensitiveContains(word) {
+                if att.name.localizedContains(word, options: query.compareOptions) {
                     wordFound = true
                     break
                 }